add cons resa all list ok

+++ src/main/java/com/epf/rentmanager/model/Client.java
@@ -19,7 +19,7 @@ public class Client {
 		
 	
 	}
-	  public Client(int id) {
+	  public Client(long id) {
 	        this.id = id;
 	    }
